Updated functions to remove dict
